refactor: extract create_rpc_client into shared utils/jsonrpc module
Deduplicates RPC client creation logic that was copied across block_info.rs,
tx_resolver.rs, and transactions.rs into a single shared function.
